Provide `Server` interface for providers to implement their language servers
This change provides the `Server` interface through which language server providers
can host their language server in one or more runtimes. The `Server` interface
exposes a numbe of features, including `LSP` and `Auth`, to be used in language
servers. This allows a portable, reusable interface for common operations and
interactions that can be hosted in a number of different runtimes, tools, and
environments.

The `standalone` runtime is provided as an initial and incomplete implementation.
It will be expanded on as future features become available. Likewise, some features
like LSP and Telemetry will be expand in future revisions.