feat: extend sendToPrompt to support optional prompt messages (#484)
## Problem
To support `aws.amazonq.ExplainIssue` in security scan for VSCode we need to have a way to send a generic message to the chat client and automatically submit that message

## Solution
Add an autoSubmit, prompt param to SendToPromptParams. This option will be disabled by default. Only when the IDE explicitily indicates that the user intends for this to automatically send will it auto submit