feat: support using rootUri for workspaceFolders and handle workspaceChangeEvent (#509)
Currently in a lot of downstream servers, the `workspace.getWorkspaceFolder` method is used for path resolution. This only relied on the `initializeParams.workspaceFolder` value which did not take into account workspace folder changes(did not listen to didChangeWorkspaceFolders) that could have happened during the session.
In addition with VS client as mentioned in aws/language-servers#1210, the workspaceFolder value is not set(more details in that PR) however it does specify that information via `rootUri` and rootPath. The utility class from that PR has now been moved centrally to this repo.
A new `getAllWorkspaceFolders()` call has been added as a common utility that servers downstream can use as single source of truth to get latest information of workspaceFolders available to the lsp server.
A follow-up PR to language-servers repo will utilize this change.