fix: remove the supports workspace check for didChangeWorkspaceFolder event (#522)
Problem

PR #509 added a check to validate whether the client supports workspaceFolders before handling the didChangeWorkspaceFolders event. This fails for VS as VS does not send the workspaceFolder value

Solution

Check has been removed as it is now assumed that whenever this is invoked clients intentionally want to send this event.
This also unblocks VS which sends the workspaceFolder information with rootUri of the initializeParams