fix: replace path.basename with custom browser-safe implementation (#636)
## Problem
`getWorkspaceFoldersFromInit` function is using the node-only `path`
module. It is imported by `base-runtime.ts` which in turn is used in
`webworker`. This causes consumers to require using polyfills, which is
not always an option as it is an unmaintained package.

https://github.com/aws/language-server-runtimes/issues/588

## Solution
Implement a custom platform-agnostic implementation of the
`path.basename` function, similar of the `joinUnixPaths` made to replace
`path.join`. This should match the behaviour of the original function
and thus not break the functionality of `getWorkspaceFoldersFromInit`

<!---
    REMINDER:
    - Read CONTRIBUTING.md first.
    - Add test coverage for your changes.
    - Link to related issues/commits.
    - Testing: how did you test your changes?
    - Screenshots if applicable
-->

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Rahmania Astrid Mochtar <rastridm@amazon.nl>