fix: add file context override for inline completion (#660)
## Problem

Support of inline completion in JL Notebook is gone since the language
server migration.
https://github.com/aws/aws-toolkit-vscode/pull/7086

## Solution

Let the IDE provide the file context. 

There is a technical limitations of the language server that it cannot
find or construct the `vscode.NotebookDocument`, `vscode.NotebookCell`.
The `workspace.getTextDocument` does not work with JL Notebook as it
gets a URI in the format of
"`vscode-notebook-cell:/../A1.ipynb#W2sZmlsZQ%3D%3D`". The language
server does not have the cell number, cell state, cursor position, etc
of the current cell.

<!---
    REMINDER:
    - Read CONTRIBUTING.md first.
    - Add test coverage for your changes.
    - Link to related issues/commits.
    - Testing: how did you test your changes?
    - Screenshots if applicable
-->

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.