chore: add timeout to ensure error flushes to stdout/err on auth failure (#589)
## Problem
No log output for initializeAuth failure

## Solution
Add a timeout before process kill to ensure the buffer flushes

Tested locally by directly modifying the prod artifact. Note that the
webpacked version appears to only output to `stdout` (instead of
`stderr`, which should be expected with a `console.error`
<!---
    REMINDER:
    - Read CONTRIBUTING.md first.
    - Add test coverage for your changes.
    - Link to related issues/commits.
    - Testing: how did you test your changes?
    - Screenshots if applicable
-->

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.